+++ app/src/main/java/com/example/hsh/homesweethome/FurnitureARActivity.java
@@ -16,7 +16,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
-import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.Model.Furniture;
 import com.google.ar.core.Anchor;
 import com.google.ar.core.HitResult;
 import com.google.ar.core.Plane;
