+++ app/src/main/java/com/example/hsh/homesweethome/FurnitureDetailsMainFragment.java
@@ -20,7 +20,7 @@ import android.widget.LinearLayout;
 import android.widget.RatingBar;
 import android.widget.TextView;
 
-import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.Model.Furniture;
 import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
