+++ app/src/main/java/com/example/hsh/homesweethome/FurnitureFragmentAdapter.java
@@ -4,10 +4,9 @@ import android.content.Context;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
-import android.util.Log;
 import android.view.ViewGroup;
 
-import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.Model.Furniture;
 
 public class FurnitureFragmentAdapter extends FragmentStatePagerAdapter {
     int PAGES_COUNT = 3;
