+++ app/src/main/java/com/example/hsh/homesweethome/Network/Interceptors/AddCookiesInterceptor.java
@@ -1,11 +1,10 @@
-package com.example.hsh.homesweethome.network.Interceptors;
+package com.example.hsh.homesweethome.Network.Interceptors;
 
 import android.content.Context;
-import android.content.pm.PackageInstaller;
 import android.util.Log;
 
 
-import com.example.hsh.homesweethome.network.Models.Session;
+import com.example.hsh.homesweethome.Network.Models.Session;
 
 import java.io.IOException;
