+++ app/src/main/java/com/example/hsh/homesweethome/Network/RetrofitClient.java
@@ -1,8 +1,5 @@
-package com.example.hsh.homesweethome.network;
+package com.example.hsh.homesweethome.Network;
 
-import com.example.hsh.homesweethome.MyApplication;
-import com.example.hsh.homesweethome.network.Interceptors.AddCookiesInterceptor;
-import com.example.hsh.homesweethome.network.Interceptors.ReceivedCookiesInterceptor;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
