+++ app/src/main/java/com/example/hsh/homesweethome/RecyclerViewAdapterMain.java
@@ -18,8 +18,8 @@ import android.widget.Filter;
 import android.widget.Filterable;
 import android.widget.TextView;
 
-import com.example.hsh.homesweethome.Models.CategoryFurniture;
-import com.example.hsh.homesweethome.Models.Furniture;
+import com.example.hsh.homesweethome.Model.CategoryFurniture;
+import com.example.hsh.homesweethome.Model.Furniture;
 
 import java.util.ArrayList;
 import java.util.Arrays;
